bossdrakor_update: drop spurious (int) cast on s16toFloat float arg; direct f32->s16 param conv matches target lwz r4 (no extsh) (99.517->99.583)

+++ src/main/dll/dll_024D_bossdrakor.c
@@ -284,7 +284,7 @@ void bossdrakor_update(int obj)
         if (((BossDrakorState*)state)->attackTimerDuration != lbl_803E6510)
         {
             s16toFloat(&((BossDrakorState*)state)->attackTimer,
-                       (int)((BossDrakorState*)state)->attackTimerDuration);
+                       ((BossDrakorState*)state)->attackTimerDuration);
         }
     }
     if ((((GameObject*)obj)->objectFlags & BOSSDRAKOR_OBJFLAG_RENDERED) == 0)
